Implemented voting

+++ app/controllers/comments_controller.rb
@@ -1,33 +1,44 @@
 class CommentsController < ApplicationController
 
- def create
-    @topic = Topic.find(params[:topic_id])
-    @post = @topic.posts.find(params[:post_id])
-    @comment = @post.comments
+  def create
+    @topic = Topic.find( params[:topic_id] )
+    @post = @topic.posts.find( params[:post_id] )
+    @comments = @post.comments
 
-    @comment = current_user.comments.build(params[:comment])
+    @comment = current_user.comments.build( comment_params )
     @comment.post = @post
-    @new_comment = Comment.new
-  
-
+    
     if @comment.save
       flash[:notice] = "Comment was created."
       redirect_to [@topic, @post]
     else
       flash[:error] = "There was an error saving the comment. Please try again."
     end
-  end
+end
 
   def destroy
     @topic = Topic.find(params[:topic_id])
     @post = @topic.posts.find(params[:post_id])
 
     @comment = @post.comments.find(params[:id])
+    authorize @comment
 
     if @comment.destroy
       flash[:notice] = "Comment was removed."
+      redirect_to [@topic, @post]
     else
       flash[:error] = "Comment couldn't be deleted. Try again."
+      redirect_to [@topic, @post]
     end
-   end
   end
+
+  private
+
+  def comment_params
+    params.require(:comment).permit(
+      :body,
+      :post_id
+    )
+  end
+
+end
